Strip email obfuscation strings
The email addresses in the HTML are translated into the correct versions
in an `onmouseover`.

The strings to substitute appear to be hard-coded, so we simply iterate
over those replacing them in turn, as per the JS function.

If these turn out to vary, we'll need to fetch and parse the string
dynamically.